move : posts dto 구조 변경
