--- apps/api-gateway/src/board/dto/create-board.dto.ts
@@ -1,17 +1,0 @@
-import { Type } from 'class-transformer';
-import { IsDate, IsNotEmpty, IsString } from 'class-validator';
-
-export class CreateBoardDto {
-  @IsString()
-  @IsNotEmpty()
-  public title: string;
-
-  @IsString()
-  @IsNotEmpty()
-  public content: string;
-
-  @Type(() => Date)
-  @IsDate()
-  @IsNotEmpty()
-  public expired_at: Date;
-}
