--- apps/api-gateway/src/board/dto/update-board.dto.ts
@@ -1,6 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateBoardDto } from './create-board.dto';
-
-export class UpdateBoardDto extends PartialType(CreateBoardDto) {
-  public id: number;
-}
