--- apps/board/src/board/dto/create-board.dto.ts
@@ -1,13 +1,0 @@
-import { IsDate, IsNotEmpty, IsString } from 'class-validator';
-
-export class CreateBoardDto {
-  @IsString()
-  public title: string;
-
-  @IsString()
-  public content: string;
-
-  @IsDate()
-  @IsNotEmpty()
-  public expired_at: Date;
-}
